+++ src/enforceViewport.ts
@@ -2,20 +2,21 @@
  * Ensure the email contains a
  * <meta name="viewport" content="width=device-width" />
  */
-function enforceViewport($: CheerioStatic) {
+function enforceViewport(
+	$: CheerioStatic,
+	desiredViewport: string = '<meta name="viewport" content="width=device-width">'
+) {
 	const viewports = $('meta[name="viewport"]');
 	const hasViewport = viewports.length > 0;
 
-	const desiredViewport = $(
-		'<meta name="viewport" content="width=device-width, initial-scale=1">'
-	);
+	const viewportElement = $(desiredViewport);
 
 	if (hasViewport) {
 		// Replace them with the one we want
 		viewports.each((index, el) => {
 			const first = index === 0;
 			if (first) {
-				$(el).replaceWith(desiredViewport);
+				$(el).replaceWith(viewportElement);
 			} else {
 				$(el).remove();
 			}
@@ -23,7 +24,7 @@ function enforceViewport($: CheerioStatic) {
 	} else {
 		// Insert a viewport
 		const head = $('head'); // Cheerio already makes sure head is present
-		head.append(desiredViewport);
+		head.append(viewportElement);
 	}
 }