+++ src/tests/enforceViewport.test.ts
@@ -21,7 +21,7 @@ describe('enforceViewport', () => {
 				<head>
 					<meta
 						name="viewport"
-						content="width=device-width, initial-scale=1"
+						content="width=device-width"
 					/>
 				</head>
 				<body>
@@ -50,7 +50,7 @@ describe('enforceViewport', () => {
 				<head>
 					<meta
 						name="viewport"
-						content="width=device-width, initial-scale=1"
+						content="width=device-width"
 					/>
 				</head>
 				<body>
@@ -77,7 +77,7 @@ describe('enforceViewport', () => {
 				<head>
 					<meta
 						name="viewport"
-						content="width=device-width, initial-scale=1"
+						content="width=device-width"
 					/>
 				</head>
 				<body>
@@ -112,7 +112,7 @@ describe('enforceViewport', () => {
 				<head>
 					<meta
 						name="viewport"
-						content="width=device-width, initial-scale=1"
+						content="width=device-width"
 					/>
 				</head>
 				<body>